removed FormDeliveryServiceConsistentHashRegexController.js and added link to documentation in traffic_portal_properties.json
