+++ traffic_portal/app/src/common/api/index.js
@@ -27,7 +27,6 @@ module.exports = angular.module('trafficPortal.api', [])
     .service('cdnService', require('./CDNService'))
     .service('changeLogService', require('./ChangeLogService'))
     .service('coordinateService', require('./CoordinateService'))
-    .service('deliveryServiceConsistentHashRegexService', require('./DeliveryServiceConsistentHashRegexService'))
     .service('deliveryServiceService', require('./DeliveryServiceService'))
     .service('deliveryServiceRegexService', require('./DeliveryServiceRegexService'))
     .service('deliveryServiceRequestService', require('./DeliveryServiceRequestService'))
