+++ traffic_portal/app/src/traffic_portal_properties.json
@@ -191,7 +191,7 @@
           },
           "consistentHashRegex": {
             "title": "Consistent Hash Regex",
-            "desc": "Regex used to extract parts of the request path using grouping in order to build a consistent request string to be used for cache selection. If this field is set on a steering delivery service, it will override the regex set on the target delivery services."
+            "desc": "Regex used to extract parts of the request path using grouping in order to build a consistent request string to be used for cache selection. If this field is set on a steering delivery service, it will override the regex set on the target delivery services.<br><br><a href=\"http://traffic-control-cdn.readthedocs.io/en/latest/admin/traffic_router.html#pattern-based-consistenthash\" target=\"_blank\">See Pattern-Based Consistent Hashing</a>"
           },
           "deepCachingType": {
             "title": "Deep Caching",
